ObjLoader by Robert Smith. Implement BVH
