+++ backup_bunny.txt
@@ -1,0 +1,13 @@
+imsize 800 600
+eye 0 1 3
+viewdir 0 0 -1
+hfov 90
+updir 0 1 0
+bkgcolor 0.356 0.698 0.976 1.0
+
+light -20 70 20 1 1 1 1
+
+
+
+
+
