+++ config.txt
@@ -1,0 +1,25 @@
+imsize 800 600
+eye 0 1 3
+viewdir 0 0 -1
+hfov 90
+updir 0 1 0
+bkgcolor 0.356 0.698 0.976 1.0
+
+light -20 70 20 1 1 1 1
+
+
+
+v -20 -20 -20
+v 20 20 -20
+v -20 20 -20
+v 20 -20 -20
+
+vt 0 1
+vt 1 0
+vt 0 0
+vt 1 1
+
+mtlcolor 0.529 0.807 0.921 0.33 0.66 0.99 0.4 0.8 0 64 1 0.2
+texture textures/harbor.ppm
+f 1/1 2/2 3/3
+f 1/1 4/4 2/2
