+++ scene.obj
@@ -1,0 +1,17 @@
+v -6 -3 -6
+v 6 -3 -15
+v -6 -3 -15
+v 6 -3 -6
+v -1 5 -9
+v 0 5 -10
+v -1 5 -10
+
+vn 0 1 0
+
+vt 0 1
+vt 1 0
+vt 0 0
+vt 1 1
+
+f 1/1/1 2/2/1 3/3/1
+f 1/1/1 4/4/1 2/2/1
