+++ x64/Debug/WhittedStyleRT.Build.CppClean.log
@@ -1,0 +1,9 @@
+d:\documents\cpp_workplace\cg\csci5607\csci5607\whittedstylert\whittedstylert\x64\debug\whittedstylert.exe
+d:\documents\cpp_workplace\cg\csci5607\csci5607\whittedstylert\whittedstylert\whittedstylert\x64\debug\whittedstylert.ilk
+d:\documents\cpp_workplace\cg\csci5607\csci5607\whittedstylert\whittedstylert\x64\debug\whittedstylert.pdb
+d:\documents\cpp_workplace\cg\csci5607\csci5607\whittedstylert\whittedstylert\whittedstylert\x64\debug\vc143.idb
+d:\documents\cpp_workplace\cg\csci5607\csci5607\whittedstylert\whittedstylert\whittedstylert\x64\debug\vc143.pdb
+d:\documents\cpp_workplace\cg\csci5607\csci5607\whittedstylert\whittedstylert\whittedstylert\x64\debug\whittedstylert.tlog\cl.command.1.tlog
+d:\documents\cpp_workplace\cg\csci5607\csci5607\whittedstylert\whittedstylert\whittedstylert\x64\debug\whittedstylert.tlog\link.command.1.tlog
+d:\documents\cpp_workplace\cg\csci5607\csci5607\whittedstylert\whittedstylert\whittedstylert\x64\debug\whittedstylert.tlog\link.read.1.tlog
+d:\documents\cpp_workplace\cg\csci5607\csci5607\whittedstylert\whittedstylert\whittedstylert\x64\debug\whittedstylert.tlog\link.write.1.tlog
