+++ x64/Debug/WhittedStyleRT.exe.recipe
@@ -1,0 +1,11 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project>
+  <ProjectOutputs>
+    <ProjectOutput>
+      <FullPath>D:\Documents\cpp_workplace\CG\CSCI5607\CSCI5607\WhittedStyleRT\WhittedStyleRT\x64\Debug\WhittedStyleRT.exe</FullPath>
+    </ProjectOutput>
+  </ProjectOutputs>
+  <ContentFiles />
+  <SatelliteDlls />
+  <NonRecipeFileRefs />
+</Project>
