+++ x64/Debug/WhittedStyleRT.tlog/WhittedStyleRT.lastbuildstate
@@ -1,0 +1,2 @@
+PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.32.31326:TargetPlatformVersion=10.0.19041.0:
+Debug|x64|D:\Documents\cpp_workplace\CG\CSCI5607\CSCI5607\WhittedStyleRT\WhittedStyleRT\|
